Added a brain called Controller class and seperated it from the Parser class. Found a bug that considers two "unnamed" modules as the same module, preventing me from adding modules that has different Id, but "unnamed" name
