+++ data/semesterList.csv
@@ -1,2 +1,3 @@
 id,name,mc,semester,isDone,grade
-cS1010,/,4,3,false,/
+cS1010,/,4,4,false,/
+cs2113,/,4,5,false,/
