+++ src/main/java/seedu/duke/data/SemModulesList.java
@@ -36,6 +36,12 @@ public class SemModulesList extends ArrayList<SelectedModule> {
 
     public boolean isInList(String moduleIdentifier) {
         for (SelectedModule module: this) {
+            if (module.getName().equals("unnamed") && moduleIdentifier.equals("unnamed")) {
+                return false;
+            }
+            if (module.getId().equals("unnamed") && moduleIdentifier.equals("unnamed")) {
+                return false;
+            }
             if (module.getName().equals(moduleIdentifier) || module.getId().equals(moduleIdentifier)) {
                 return true;
             }
